bitcoin-transactions/requirements.txt
requests
requests[socks]

mnemonic==0.21
pycryptodome>=3.17
coincurve>=18.0.0
python-bitcoinlib>=0.12.2

cryptography>=40.0
base58>=2.1.1

web3
web3[tester,eth-abi]






